+++ components/SearchBox.vue
@@ -55,32 +55,32 @@ export default {
 
 .vue-component.search-box {
     margin: 0 1px 1em 0;
-    position: relative;
+	display: flex;
+	align-items: stretch;
 
 	input, .icon {
-		height: 1.5em;
 		font-size: 1em;
 		margin: 0;
 	}
 	input {
-		padding: 0.25em 0.3em;
+		flex-grow: 1;
+		padding: 0.3em;
 		padding-left: 1.9em;
 		z-index: 1;
-		display: inline-block;
 		border: 1px solid #ccc;
 		box-sizing: content-box;
 		background-color: #fff;
-		width: calc(100% - 1.9em - 0.25em - 2px);
+		box-sizing: border-box;
+		width: 100%;
+		margin-left: -1.3em;
 	}
 	.icon {
+		display: flex;
+		align-items: center;
 		user-select: none;
-		margin-top: 0.3em;
 		margin-left: 0.3em;
 		width: 1em;
 		z-index: 2;
-		position: absolute;
-		top: 0;
-		left: 0;
 	}
 	&.compact {
 		font-size: 0.9em;